chore: update build scripts

+++ deploy.sh
@@ -6,22 +6,29 @@ appName=get-sub
 user=root
 path=/var/www/
 
-fs-ops zip dist -o=dist/${appName}.zip
-
 scp dist/${appName}.zip ${user}@${ip}:${path}
 ssh ${user}@${ip} bash -c "'
 
 cd ${path}
 
-mkdir -p ${appName}/node_modules
-rm -rf ${appName}-bak
-mv -f ${appName} ${appName}-bak
+if [ -d ${appName} ]; then
+  rm -rf ${appName}-bak
+  mv -f ${appName} ${appName}-bak
+fi
 
 unzip -o ${appName}.zip -d ${appName}
-mv ${appName}-bak/node_modules ${appName}
+
+if [ -d ${appName}-bak ]; then
+  for file in node_modules package-lock.json
+  do
+    if [ -f ${appName}-bak/\${file} ] || [ -d ${appName}-bak/\${file} ]; then
+      mv ${appName}-bak/\${file} ${appName}/\${file}
+    fi
+  done
+fi
 
 cd ${appName}
 npm run start:pm2
 
+exit
 '"
-
